Adding option in setting to enable and disable notifications

+++ static/app.js
@@ -3,12 +3,24 @@ var notificationArray = [];
 document.addEventListener("DOMContentLoaded", function(event) {
     console.log('Document fully loaded');
 
+    const $notificationCheckbox = $(".notificationCheckbox");
+
     var socket = io.connect(location.protocol + '//' + document.domain + ':' + location.port + '/price');
     console.log(currentDate(new Date()) + ' Socket connected!');
 
-    socket.on('connect', function() {
-        // Emitting a message to the server that the client is connected.
-        socket.emit("client_message", {data: currentDate(new Date()) + ' I´m connected with the server!'});
+     // Setting up the local storage
+
+    // If notification storage value is not defined, set it by default to true
+    if (storageAvailable('localStorage')) {
+
+        // Setting up the application if it is the first time it was opened.
+        if(localStorage.getItem("showNotifications") === null){
+            localStorage.setItem("showNotifications", "false");
+            $notificationCheckbox.prop("checked", false);
+        }
+
+    }
+    else {
 
         // Creating a random ID for the new created entry in the list to identify the created entry later.
         var milliseconds = new Date().getMilliseconds();
@@ -19,11 +31,10 @@ document.addEventListener("DOMContentLoaded", function(event) {
         var notification = document.createElement("div");
 
         // Content of the notification
-        var infoNotificationTemplate =
+        var errorNotificationTemplate =
             "        <div class='content'>" +
-            "            Connected to the server!" +
-            "            <button id='close_notification' onclick='closeNotification(\"" + newIdName + "\")'>X</button><br>" +
-            "            <small>" + currentDate(new Date()) + "</small>" +
+            "            <button id='close_notification' onclick='closeNotification(\"" + newIdName + "\")'>X</button>" +
+            "            Cannot access the local storage in your browser!" +
             "        </div>";
 
         // Accessing the notification list (<ul>)
@@ -38,8 +49,8 @@ document.addEventListener("DOMContentLoaded", function(event) {
 
         // Setting new data
         entry.id = newIdName;
-        notification.className = "notification success_notification";
-        notification.innerHTML = infoNotificationTemplate;
+        notification.className = "notification error_notification";
+        notification.innerHTML = errorNotificationTemplate;
 
         // Push the new ID name to the notification array
         notificationArray.push(newIdName);
@@ -47,6 +58,61 @@ document.addEventListener("DOMContentLoaded", function(event) {
         // Showing the amount of currently visible notifications
         var amountNotificationsElement = document.getElementById("deleteAllNotifications").getElementsByTagName("span")[0];
         amountNotificationsElement.innerText = notificationArray.length;
+
+    }
+
+    // Setting up application with the values given by the application first start or by the user by changing
+    // the values in the settings window.
+    if(localStorage.getItem("showNotifications") === "true"){
+        $notificationCheckbox.prop("checked", true);
+    }else{
+        $notificationCheckbox.prop("checked", false);
+    }
+
+    // "Connect" socket event
+    socket.on('connect', function() {
+        // Emitting a message to the server that the client is connected.
+        socket.emit("client_message", {data: currentDate(new Date()) + ' I´m connected with the server!'});
+
+        if(localStorage.getItem("showNotifications") === "true") {
+            // Creating a random ID for the new created entry in the list to identify the created entry later.
+            var milliseconds = new Date().getMilliseconds();
+            var randomNumber = Math.floor(Math.random() * 1000);
+            var newIdName = currentDate(new Date()).replace(" ", "-") + "-" + milliseconds + randomNumber + "-notification";
+
+            // Creating the notification
+            var notification = document.createElement("div");
+
+            // Content of the notification
+            var infoNotificationTemplate =
+                "        <div class='content'>" +
+                "            Connected to the server!" +
+                "            <button id='close_notification' onclick='closeNotification(\"" + newIdName + "\")'>X</button><br>" +
+                "            <small>" + currentDate(new Date()) + "</small>" +
+                "        </div>";
+
+            // Accessing the notification list (<ul>)
+            var notificationList = document.getElementById("notification_list");
+
+            // Creating the element
+            var entry = document.createElement('li');
+
+            // Append the new elements to the list
+            notificationList.appendChild(entry);
+            entry.appendChild(notification);
+
+            // Setting new data
+            entry.id = newIdName;
+            notification.className = "notification success_notification";
+            notification.innerHTML = infoNotificationTemplate;
+
+            // Push the new ID name to the notification array
+            notificationArray.push(newIdName);
+
+            // Showing the amount of currently visible notifications
+            var amountNotificationsElement = document.getElementById("deleteAllNotifications").getElementsByTagName("span")[0];
+            amountNotificationsElement.innerText = notificationArray.length;
+        }
 
     });
 
@@ -108,47 +174,50 @@ document.addEventListener("DOMContentLoaded", function(event) {
             }
         }
 
-        // Creating a random ID for the new created entry in the list to identify the created entry later.
-        var milliseconds = new Date().getMilliseconds();
-        var randomNumber = Math.floor(Math.random() * 1000);
-        var newIdName = currentDate(new Date()).replace(" ", "-") + "-" + milliseconds + randomNumber + "-notification";
+        if(localStorage.getItem("showNotifications") === "true") {
+            // Creating a random ID for the new created entry in the list to identify the created entry later.
+            var milliseconds = new Date().getMilliseconds();
+            var randomNumber = Math.floor(Math.random() * 1000);
+            var newIdName = currentDate(new Date()).replace(" ", "-") + "-" + milliseconds + randomNumber + "-notification";
 
-        // Creating the notification
-        var notification = document.createElement("div");
+            // Creating the notification
+            var notification = document.createElement("div");
 
-        // Content of the notification
-        var infoNotificationTemplate =
-            "        <div class='content'>" +
-            "            Price updated to: <u>" + msg.price + "</u>" +
-            "            <button id='close_notification' onclick='closeNotification(\"" + newIdName + "\")'>X</button><br>" +
-            "            <small>" + currentDate(new Date()) + "</small>" +
-            "        </div>";
+            // Content of the notification
+            var infoNotificationTemplate =
+                "        <div class='content'>" +
+                "            Price updated to: <u>" + msg.price + "</u>" +
+                "            <button id='close_notification' onclick='closeNotification(\"" + newIdName + "\")'>X</button><br>" +
+                "            <small>" + currentDate(new Date()) + "</small>" +
+                "        </div>";
 
-        // Accessing the notification list (<ul>)
-        var notificationList = document.getElementById("notification_list");
+            // Accessing the notification list (<ul>)
+            var notificationList = document.getElementById("notification_list");
 
-        // Creating the element
-        var entry = document.createElement('li');
-        notificationList.appendChild(entry);
+            // Creating the element
+            var entry = document.createElement('li');
+            notificationList.appendChild(entry);
 
-        // Append the new elements to the list
-        entry.appendChild(notification);
+            // Append the new elements to the list
+            entry.appendChild(notification);
 
-        // Setting new data
-        entry.id = newIdName;
-        notification.className = "notification info_notification";
-        notification.innerHTML = infoNotificationTemplate;
+            // Setting new data
+            entry.id = newIdName;
+            notification.className = "notification info_notification";
+            notification.innerHTML = infoNotificationTemplate;
 
-        // Push the new ID name to the notification array
-        notificationArray.push(newIdName);
+            // Push the new ID name to the notification array
+            notificationArray.push(newIdName);
 
-        // Showing the amount of currently visible notifications
-        var amountNotificationsElement = document.getElementById("deleteAllNotifications").getElementsByTagName("span")[0];
-        amountNotificationsElement.innerText = notificationArray.length;
+            // Showing the amount of currently visible notifications
+            var amountNotificationsElement = document.getElementById("deleteAllNotifications").getElementsByTagName("span")[0];
+            amountNotificationsElement.innerText = notificationArray.length;
+        }
 
     });
 
     // Error in very rare moments
+    // The settings for the notifications won´t play a role here because this is a error. Period.
     socket.on('error', function (error) {
 
         // Creating a random ID for the new created entry in the list to identify the created entry later.
@@ -190,7 +259,7 @@ document.addEventListener("DOMContentLoaded", function(event) {
 
     });
 
-    console.log(socket);
+    // console.log(socket);
 
     // Check if the close button for the settings dialog window was pressed (event handler)
     var closeButton = document.getElementById('closeSettingsDialogWindow');
@@ -198,6 +267,10 @@ document.addEventListener("DOMContentLoaded", function(event) {
         var settingsDialogWindow = document.getElementById("settingsDialogWindow");
         settingsDialogWindow.style.display = "none";
     };
+
+    $notificationCheckbox.click(function () {
+        localStorage.setItem("showNotifications", this.checked.toString());
+    });
 
 });
 
@@ -212,7 +285,7 @@ function addZero(i) {
 // Function which returns the current date in a more readable string
 function currentDate(/** Date */date) {
     var day = date.getDate();
-    var month = date.getMonth();
+    var month = date.getMonth() + 1;
     var year = date.getFullYear();
 
     var hour = addZero(date.getHours());
@@ -256,5 +329,30 @@ function deleteAllNotifications() {
 function openSettingsDialog() {
     var settingsDialogWindow = document.getElementById("settingsDialogWindow");
     settingsDialogWindow.style.display = "block";
+}
+
+// Check if the a storage is available
+function storageAvailable(type) {
+    try {
+        var storage = window[type],
+            x = '__storage_test__';
+        storage.setItem(x, x);
+        storage.removeItem(x);
+        return true;
+    }
+    catch(e) {
+        return e instanceof DOMException && (
+            // everything except Firefox
+            e.code === 22 ||
+            // Firefox
+            e.code === 1014 ||
+            // test name field too, because code might not be present
+            // everything except Firefox
+            e.name === 'QuotaExceededError' ||
+            // Firefox
+            e.name === 'NS_ERROR_DOM_QUOTA_REACHED') &&
+            // acknowledge QuotaExceededError only if there's something already stored
+            storage.length !== 0;
+    }
 }
